Added threshold when comparing historic and current route
Checks the percentage of changes that have occurred between two
traceroute paths and if it exceeds the threshold, it will setup the
email notification message that will be sent out at the end
of the script. The reason for this inclusion is to view only significant
changes that happen since the last test ran.